Creact ProcessButton component
